% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{abs}
\alias{abs}
\title{apply the abs function on all the timeseries contained in this Dataset}
\usage{
abs(x)
}
\arguments{
\item{x}{a Dataset where we want to apply the `abs`}
}
\value{
a Dataset with all the timeseries with the `abs` applied
}
\description{
apply the abs function on all the timeseries contained in this Dataset
}
