% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{abs_ds}
\alias{abs_ds}
\title{apply the abs function on all the timeseries contained in this Dataset}
\usage{
abs_ds(x)
}
\arguments{
\item{x}{a Dataset we want to apply the abs}
}
\value{
a Dataset with all the timesereis with the abs applied
}
\description{
apply the abs function on all the timeseries contained in this Dataset
}

