% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.r
\name{annual}
\alias{annual}
\title{Esegue l'annual del dataset}
\usage{
annual(x)
}
\arguments{
\item{x}{Dataset da annualizzare}
}
\value{
un dataset annualizzato
}
\description{
Se le serie contenute sono gia' annuali, non fa nulla e
lascia la serie cosi' com'e'
}
