% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataset.r
\docType{methods}
\name{as.dataset}
\alias{as.dataset}
\alias{as.dataset,list-method}
\alias{as.dataset,Dataset-method}
\alias{as.dataset,data.frame-method}
\title{Casts to a Dataset}
\usage{
as.dataset(x, ...)

\S4method{as.dataset}{list}(x)

\S4method{as.dataset}{Dataset}(x)

\S4method{as.dataset}{data.frame}(x)
}
\arguments{
\item{x}{a generico object with "[[" method defined and names}

\item{...}{forza la creazione di un nuovo dataset, anche se x e' un `Dataset`}
}
\value{
a Dataset with data defined in x
}
\description{
Casts to a Dataset
}
