% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{as.dataset}
\alias{as.dataset}
\title{Casts to a Dataset}
\usage{
as.dataset(x)
}
\arguments{
\item{x}{a generico object with "[[" method defined and names}

\item{...}{forza la creazione di un nuovo dataset, anche se x e' un `Dataset`}
}
\value{
a Dataset with data defined in x
}
\description{
Casts to a Dataset
}
