% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{as.list.Dataset}
\alias{as.list.Dataset}
\title{Ritorna il `Dataset` come `list` di oggetti}
\usage{
as.list.Dataset(x)
}
\arguments{
\item{x}{un generico oggetto R}

\item{...}{generics passati a base::as.list}
}
\value{
un `Dataset`
}
\description{
Ritorna il `Dataset` come `list` di oggetti
}
\seealso{
base::as.list
}
