% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{as.list}
\alias{as.list}
\alias{as.list.Dataset}
\title{Ritorna il `Dataset` come `list` di oggetti}
\usage{
\method{as.list}{Dataset}(x, ...)
}
\value{
una `list` di Timeseries
}
\description{
Ritorna il `Dataset` come `list` di oggetti
}
\seealso{
base::as.list
}
