% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{as.vector}
\alias{as.vector}
\title{Elenco nomi Dataset}
\usage{
\S4method{as.vector}{Dataset}(x)
}
\arguments{
\item{x}{un istanza di Dataset}
}
\value{
Una rappresentazione a lista del Dataset
}
\description{
Ritorna un elenco di nomi di serie storiche serie storiche contenuti
nel Dataset
}
\seealso{
\code{link{rcf::names}}
}
