% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{$,Dataset-method}
\alias{$,Dataset-method}
\title{Accesses the timeseries/object based on its name}
\usage{
\S4method{$}{Dataset}(x, name)
}
\arguments{
\item{x}{dataset}

\item{name}{nome dell'oggetto da estrarre}
}
\description{
Accesses the timeseries/object based on its name
}

