% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_merging.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine two datasets}
\usage{
combine_datasets(
  a,
  b,
  strict = c("equal"),
  left_name = "left",
  right_name = "right"
)
}
\arguments{
\item{a}{A dataset object.}

\item{b}{A dataset object.}

\item{strict}{Strictness of ID matching.}

\item{left_name}{Name for left values.}

\item{right_name}{Name for right values.}
}
\value{
A dataframe in the internal \code{"scuffed_long"} format.
}
\description{
Joins the long forms of two datasets into a single intermediate table.
}
