% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.r
\docType{methods}
\name{copy}
\alias{copy}
\alias{copy,Dataset-method}
\title{Makes a deep copy of the object}
\usage{
copy(x)

\S4method{copy}{Dataset}(x)
}
\arguments{
\item{x}{oggetto da copiare}
}
\value{
una copia di `x`
}
\description{
Makes a deep copy of the object
}
