% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.r
\name{copy}
\alias{copy}
\title{Questa funzione esegue una deep copy dell'oggetto}
\usage{
copy(x)
}
\arguments{
\item{x}{oggetto da copiare}
}
\value{
una copia di `x`
}
\description{
Questa funzione esegue una deep copy dell'oggetto
}
