% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\alias{Dataset}
\title{Classe contenitore di dati (serie storiche)}
\description{
Classe contenitore di dati (serie storiche)
}
\section{Slots}{

\describe{
\item{\code{data}}{hash::hash containing data}

\item{\code{url}}{path where data is}
}}

