% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{class}
\name{Dataset}
\alias{Dataset}
\alias{Dataset-class}
\title{Dataset OOP}
\usage{
Dataset(url, ids)
}
\arguments{
\item{url}{path where data is}

\item{ids}{only IDs area leaded in this Dataset (character array of names)}
}
\description{
Classe contenitore di dati (serie storiche)
}
\author{
Giuseppe Acito
}

