% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{dataset}
\alias{dataset}
\title{Dataset OOP}
\usage{
dataset(...)
}
\arguments{
\item{...}{Qui si puo specificare
- una stringa con un URL da cui costruire il Dataset
  (Grafo, Speakeasy library, Speakeasy List)
- una lista di oggetti}
}
\value{
Il Dataset richiesto.
}
\description{
Costruisce un dataset.
}
\examples{
\dontrun{
ds <-Dataset()
ds <- Dataset('/path/to/something/useful')
given \\code{lll} is a list of timeseries
ds <- dataset(lll)
}
}
\author{
Giuseppe Acito
}

