% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{dataset_build}
\alias{dataset_build}
\title{Create a dataset}
\usage{
dataset_build(df, ids)
}
\arguments{
\item{df}{A data frame containing identifier and value columns.}

\item{ids}{Character vector of column names that define the unique row
identifier(s). Must not include the future value axis column.}
}
\value{
A dataset object (in wide format).
}
\description{
Builds a dataset object from a data frame and a vector of identifier
column names. The resulting dataset is stored in wide format by default.
}
