% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{dataset_collapse}
\alias{dataset_collapse}
\title{Collapse a dataset}
\usage{
dataset_collapse(dataset)
}
\arguments{
\item{dataset}{A dataset object.}
}
\value{
A dataset with empty rows and columns removed.
}
\description{
Removes rows and columns that contain only missing values (\code{NA}). This
is a space-saving operation that does not change the logical content of
the dataset.
}
