% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_arithmetic.R, R/set_operands.R
\name{dataset_intersect}
\alias{dataset_intersect}
\alias{>.dataset}
\title{Set intersection (a & b)}
\usage{
dataset_intersect(a, b)

\method{>}{dataset}(a, b)
}
\arguments{
\item{a}{A dataset object.}

\item{b}{A dataset object.}
}
\value{
A dataset with the intersection values taken from \code{a}.
}
\description{
Keeps cells from \code{a} that also exist in \code{b}.
}
