% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{date_index}
\alias{date_index}
\title{ritorna un array di date da un index di xts}
\usage{
date_index(idx, freq)
}
\arguments{
\item{idx}{vettore di periodi/date (ottenuto con zoo::index)}

\item{freq}{frequenza del vettore}
}
\value{
un vettore di date, con correzzione se frequenza annuale
}
\description{
la funzione ha cura di correggere il problema delle date annuali di xts
}
