% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djoin.r
\docType{methods}
\name{djoin}
\alias{djoin}
\alias{djoin,Dataset,Dataset-method}
\alias{djoin,Dataset,Dataset,ANY-method}
\title{Effettua il join ad una data per due Dataset}
\usage{
djoin(historic, x, date)

\S4method{djoin}{Dataset,Dataset}(historic, x, date)
}
\arguments{
\item{historic}{il dataset storico}

\item{x}{il dataset nuovo}

\item{date}{una periodo di join}
}
\value{
Un dataset joinato
}
\description{
Joina i due dataset \code{x} e \code{historic} alla data (\code{date})
specificata.
}
