% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{empty_set}
\alias{empty_set}
\title{Create an empty dataset}
\usage{
empty_set()
}
\value{
An empty dataset object (class \code{"dataset"}, state \code{"wide"}).
}
\description{
Returns a dataset with zero rows and zero ID columns.
}
