% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataset.r
\name{from_data_frame_to_ts}
\alias{from_data_frame_to_ts}
\title{A converter from a data.frame to ts}
\usage{
from_data_frame_to_ts(y)
}
\arguments{
\item{y}{the data.frame to be converted}
}
\value{
a ts object
}
\description{
The data.frame *must* follow a fixed schema (columns):
- idx with obervations timestamps
- year
- period
- freq
- obs (the observations' values)
}
