% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{from_json}
\alias{from_json}
\title{converts a JSON timeseries to a BIMETS timeseries timeseries}
\usage{
from_json(json)
}
\arguments{
\item{json}{a character array containing a json string}
}
\value{
a BIMETS timeseries
}
\description{
converts a JSON timeseries to a BIMETS timeseries timeseries
}
\seealso{
\code{from_list}
}
