% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{from_list}
\alias{from_list}
\title{converts a list structure to a BIMETS timeseries}
\usage{
from_list(lista)
}
\arguments{
\item{lista}{a list object}
}
\value{
a timeseries
}
\description{
converts a list structure to a BIMETS timeseries
}

