% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\docType{methods}
\name{fullSummary}
\alias{fullSummary}
\alias{fullSummary,Dataset-method}
\title{tabs all the series in the Dataset}
\usage{
fullSummary(ds)

\S4method{fullSummary}{Dataset}(ds)
}
\arguments{
\item{ds}{a `Dataset` instance}
}
\description{
tabs all the series in the Dataset
}
