% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_utils.R
\name{intersect_with}
\alias{intersect_with}
\title{Intersect with another dataset}
\usage{
intersect_with(a, b)
}
\arguments{
\item{a}{A dataset object.}

\item{b}{A dataset object.}
}
\value{
A dataset containing intersection values from \code{a}.
}
\description{
Keeps only the cells in \code{a} for which there is a corresponding non‑\code{NA}
cell in \code{b}. Uses left precedence and equal or greater ID requirement.
}
