% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is.dataset}
\alias{is.dataset}
\title{checks if \code{x} is an object of type \code{Dataset}}
\usage{
is.dataset(x)
}
\arguments{
\item{x}{a generic object}
}
\value{
\code{TRUE} if \code{x} is a \code{Dataset}, \code{FALSE} otherwise
}
\description{
checks if \code{x} is an object of type \code{Dataset}
}
