% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_csv_library}
\alias{is_csv_library}
\title{Controlla che `path` sia una library CSV}
\usage{
is_csv_library(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
`TRUE` se e' una library CSV, `FALSE` otherwise
}
\description{
Controlla che `path` sia una library CSV
}

