% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_grafo}
\alias{is_grafo}
\title{Controlla che il percorso \code{path} contenga una struttura a
Grafo sul file system}
\usage{
is_grafo(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
\code{TRUE} se il percorso e' un grafo valido,
        altrimenti \code{FALSE}
}
\description{
Controlla che il percorso \code{path} contenga una struttura a
Grafo sul file system
}
