% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_JSON}
\alias{is_JSON}
\title{Controlla che il percorso \code{path} e' un file JSON}
\usage{
is_JSON(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
\code{TRUE} se il percorso e' un file JSON valido,
        altrimenti \code{FALSE}
}
\description{
Controlla che il percorso \code{path} e' un file JSON
}
