% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_speakeasy}
\alias{is_speakeasy}
\title{Controlla che il percorso \code{path} e' un contenitore di dati Speakeasy}
\usage{
is_speakeasy(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
\code{TRUE} se il percorso e' una library o una list Speakeasy,
        altrimenti \code{FALSE}
}
\description{
Controlla che il percorso \code{path} e' un contenitore di dati Speakeasy
}
\author{
Giuseppe Acito
}
\seealso{
is_speakeas_list is_speakeasy_library
}

