% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_speakeasy_library}
\alias{is_speakeasy_library}
\title{Controlla che il percorso \code{path} e' una library Speakeasy}
\usage{
is_speakeasy_library(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
\code{TRUE} se il percorso e' una library Speakeasy,
        altrimenti \code{FALSE}
}
\description{
Controlla che il percorso \code{path} e' una library Speakeasy
}
\author{
Giuseppe Acito
}

