% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{is_speakeasy_list}
\alias{is_speakeasy_list}
\title{Controlla che il percorso \code{path} e' un file list Speakeasy}
\usage{
is_speakeasy_list(path)
}
\arguments{
\item{path}{percorso da controllare}
}
\value{
\code{TRUE} se il percorso e' una list Speakeasy,
        altrimenti \code{FALSE}
}
\description{
Controlla che il percorso \code{path} e' un file list Speakeasy
}
\author{
Giuseppe Acito
}

