% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{length}
\alias{length}
\alias{length,Dataset-method}
\title{Returns the length of the object}
\usage{
\S4method{length}{Dataset}(x)
}
\arguments{
\item{x}{object to retrieve the length}
}
\value{
the length of the object x
}
\description{
Returns the length of the object
}
