% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{length}
\alias{length}
\title{Elenco nomi Dataset}
\usage{
\S4method{length}{Dataset}(x)
}
\arguments{
\item{x}{un istanza di Dataset}
}
\value{
il numero di serie storiche nel Dataset
}
\description{
Ritorna il numero di oggetti contenute nel Dataset
}
\author{
Giuseppe Acito
}

