% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{load_dataset_csv}
\alias{load_dataset_csv}
\title{internal functions}
\usage{
load_dataset_csv(path, ids = NULL)
}
\arguments{
\item{path}{path where CSV files are}

\item{ids}{names of the series to load}
}
\value{
a named `list` with timeseries
}
\description{
load a list of timeseries placed in path as CSV files.
}
\note{
the path is scanned recursively
}
\seealso{
tsRead_nativo
}

