% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{load_dataset_grafo}
\alias{load_dataset_grafo}
\title{Internal Functions}
\usage{
load_dataset_grafo(path)
}
\arguments{
\item{path}{percorso da cui caricare il Dataset}
}
\value{
un `Dataset`
}
\description{
Carica un Dataset da un percorso di Grafo
}
\note{
Funzione interna, potrebbe cambiare
}

