% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{load_dataset_json}
\alias{load_dataset_json}
\title{Internal Functions}
\usage{
load_dataset_json(path)
}
\arguments{
\item{path}{percorso da cui caricare il Dataset}
}
\value{
un `Dataset`
}
\description{
Carica un `Dataset` a partire da un file JSON
}
\note{
Funzione interna, potrebbe cambiare
}
