% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{merge.Dataset}
\alias{merge.Dataset}
\title{merges two timeseries datasets}
\usage{
\method{merge}{Dataset}(x, y, ...)
}
\arguments{
\item{x}{first dataset}

\item{y}{second dataset}

\item{...}{just for compliance}
}
\value{
a dataset with the timeseries merged
}
\description{
merges two timeseries datasets
}
\note{
raises a warning for timeseries names not common in `x` and `y`
}
