% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{merge}
\alias{merge}
\title{merges two timeseries datasets}
\usage{
merge(x,y)
}
\arguments{
\item{x}{first dataset}

\item{y}{second dataset}
}
\value{
a dataset with the timeseries merged
}
\description{
merges two timeseries datasets
}
\note{
raises a warning for timeseries names not common in `x` and `y`
}
