% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_merging.R
\name{merge_func}
\alias{merge_func}
\title{Build a merging function}
\usage{
merge_func(
  op = c("left", "right", "setdiff", "symdiff", "intersect", "union"),
  prc = c("left", "right")
)
}
\arguments{
\item{op}{Operation name.}

\item{prc}{Precedence rule.}
}
\value{
A function that takes two value vectors and returns the merged
vector.
}
\description{
Creates the value-level logic for a merge operation.
}
