% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{names}
\alias{names}
\alias{names,Dataset-method}
\title{Yields the names contained in the Object}
\usage{
\S4method{names}{Dataset}(x)
}
\arguments{
\item{x}{object where "names" are from}
}
\description{
Yields the names contained in the Object
}
