% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{names}
\alias{names}
\title{Elenco nomi Dataset}
\usage{
\S4method{names}{Dataset}(x)
}
\arguments{
\item{x}{un istanza di Dataset}
}
\value{
un character array contenente i nomi delle serie storiche nel Dataset
}
\description{
Elenco dei nomi di serie contenuti nel Dataset
}
\author{
Giuseppe Acito
}

