% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{print.dataset}
\alias{print.dataset}
\title{Print a dataset}
\usage{
\method{print}{dataset}(x, ...)
}
\arguments{
\item{x}{A dataset object.}

\item{...}{Further arguments passed to other print methods.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Provides a summary of the dataset's state, ID columns, value columns,
and x-axis, then delegates to the next print method to show the actual
data.
}
