% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{"-"}
\alias{"-"}
\title{Differenza tra Dataset}
\usage{
\S4method{-}{Dataset,Dataset}(e1, e2)
}
\arguments{
\item{e1}{Dataset (primo operando)}

\item{e2}{Dataset (secondo operando)}
}
\value{
Il dataset con le differenze
}
\description{
Esegue la differenza tra due dataset di oggetti (a patto che la differenza
sia definita per gli oggetti)
Se un oggetto non e' comune ai due dataset viene lanciato un warning
}

