% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{"[[<-"}
\alias{"[[<-"}
\title{Subsetting}
\usage{
\S4method{[[}{Dataset,ANY,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{un istanza di Dataset}

\item{i}{una stringa (il nome della serie storica)}

\item{j}{una stringa (never userd)}

\item{...}{God knows what is this good for}

\item{value}{l'oggetto da impostare}
}
\description{
Imposta, data una stringa ed un oggetto xts, una serie storica nel Dataset
}
\author{
Giuseppe Acito
}
