% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_diffable_text.R
\name{read_dataset}
\alias{read_dataset}
\title{Read Dataset from JSON File}
\usage{
read_dataset(path)
}
\arguments{
\item{path}{File path to the JSON file.}
}
\value{
A dataset object.
}
\description{
Reads a dataset object from a JSON file created by \code{write_dataset()}.
}
\examples{
\dontrun{
ds <- read_dataset("output.json")
}
}
