% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.r
\docType{methods}
\name{==}
\alias{==}
\alias{==,Dataset,numeric-method}
\alias{>}
\alias{>,Dataset,numeric-method}
\alias{<}
\alias{<,Dataset,numeric-method}
\alias{<=}
\alias{<=,Dataset,numeric-method}
\alias{>=}
\alias{>=,Dataset,numeric-method}
\title{equals relation operator}
\usage{
\S4method{==}{Dataset,numeric}(e1, e2)

\S4method{>}{Dataset,numeric}(e1, e2)

\S4method{<}{Dataset,numeric}(e1, e2)

\S4method{<=}{Dataset,numeric}(e1, e2)

\S4method{>=}{Dataset,numeric}(e1, e2)
}
\arguments{
\item{e1}{Dataset}

\item{e2}{numeric}
}
\description{
equals relation operator

greater-than relation operator

less-than relation operator

less-or-equal relation operator

greater-or-equal relation operator
}
