% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.r
\docType{methods}
\name{round}
\alias{round}
\alias{round,Dataset-method}
\alias{round,Dataset,ANY-method}
\title{Yields a rounding of the object}
\usage{
\S4method{round}{Dataset}(x, digits = 0)
}
\arguments{
\item{x}{object to round}

\item{digits}{numeber of digits for roundings}
}
\value{
the same object, with roundings
}
\description{
Yields a rounding of the object
}
