% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{save.dataset}
\alias{save.dataset}
\title{Internal Functions}
\usage{
save.dataset(data, path)
}
\arguments{
\item{data}{il `Dataset` da salvare}

\item{path}{percorso del file JSON su cui salvare il `Dataset`}
}
\description{
Salva un dataset come file JSON
}
\note{
Funzione interna, potrebbe cambiare
}

