% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{saveDataset}
\alias{saveDataset}
\title{Salva un Dataset}
\usage{
\S4method{saveDataset}{Dataset,character}(x, path)
}
\arguments{
\item{x}{un Dataset}

\item{path}{percorso dove salvare il Dataset}
}
\description{
Salva un Dataset.
}
\seealso{
\code{saveGraph}
}
\author{
Giuseppe Acito
}
