% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{set_attr}
\alias{set_attr}
\title{Set attributes for a dataset}
\usage{
set_attr(df, ids, x_axis, state)
}
\arguments{
\item{df}{A dataframe.}

\item{ids}{ID columns.}

\item{x_axis}{X-axis column.}

\item{state}{Dataset state.}
}
\description{
Set attributes for a dataset
}
