% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{shortSummary}
\alias{shortSummary}
\title{Print a summary of this Dataset}
\usage{
shortSummary(object)
}
\arguments{
\item{object}{a `Dataset` object}
}
\description{
For each series prints it's stard period, end period and frequency
}
