% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{show}
\alias{show}
\alias{show,Dataset-method}
\title{shows the object on the stdout}
\usage{
\S4method{show}{Dataset}(object)
}
\arguments{
\item{object}{object to printout}
}
\description{
shows the object on the stdout
}
