% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{show}
\alias{show}
\title{Show Method}
\usage{
\S4method{show}{Dataset}(object)
}
\arguments{
\item{x}{Dataset da mostrare}
}
\description{
Default metodo show per la classe Dataset
}
\author{
Giuseppe Acito
}
\seealso{
\code{Dataset}
}

