% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{stretch}
\alias{stretch}
\title{Allunga tutte le serie del Dataset.}
\usage{
stretch(x, periodo)
}
\arguments{
\item{x}{nome del Dataset}

\item{periodo}{coppia `(anno, trimestre)` rispetto cui allungare le serie}
}
\value{
il dataset con le serie allungate
}
\description{
La funzione segue la seguente convezione:
   - se la serie termina in `C` o `S` e' una condistenza e va allungata
     con il medesimo valore
   - altrimenti e' un flusso e va allungata con zero
}
