% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\docType{methods}
\name{[}
\alias{[}
\alias{[,Dataset,ANY,missing,ANY-method}
\alias{[[}
\alias{[[,Dataset,character-method}
\alias{[[,Dataset,ANY,missing,ANY-method}
\alias{[<-}
\alias{[<-,Dataset,ANY,missing,ANY-method}
\alias{[[<-}
\alias{[[<-,Dataset,ANY,missing,ANY-method}
\title{standard method for accessing indexed elements in a container}
\usage{
\S4method{[}{Dataset,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{Dataset,character}(x, i)

\S4method{[}{Dataset,ANY,missing,ANY}(x, i, j, ...) <- value

\S4method{[[}{Dataset,ANY,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Dataset}

\item{i}{index element identifier}

\item{j}{dunno what to do with this}

\item{...}{just for compliance}

\item{drop}{just for compliance}

\item{value}{the value to be set for `i`}
}
\description{
standard method for accessing indexed elements in a container

standard method for accessing indexed elements in a container

standard method for accessing indexed elements in a container

standard method for accessing indexed elements in a container
}
