% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{"["}
\alias{"["}
\alias{"[["}
\alias{"[<-"}
\title{Subsetting}
\usage{
\S4method{[}{Dataset,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{Dataset,character}(x, i)

\S4method{[}{Dataset,ANY,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{un istanza di Dataset}

\item{i}{una stringa (il nome della serie storica)}

\item{j}{una stringa (never userd)}

\item{...}{God knows what is this good for}

\item{drop}{Same as before}

\item{value}{l'oggetto da impostare}
}
\value{
Ritorna un sub-Dataset

Ritorna una serie storica on una list di serie storiche
}
\description{
Ritorna una singola serie storica o un sub Dataset di serie storiche

Ritorna un list o una singola  serie storica

Imposta, data una stringa ed un oggetto xts, una serie storica nel Dataset
}
