% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.r
\name{sum.Dataset}
\alias{sum.Dataset}
\title{Esegue il `sum` sul Dataset}
\usage{
\method{sum}{Dataset}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{dataset da sommare}

\item{...}{other objects to add}

\item{na.rm}{Not used, just to be compliant with sum}
}
\description{
Esegue il `sum` sul Dataset
}
