% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{to_json}
\alias{to_json}
\title{Converts to JSON a timeseries}
\usage{
to_json(t)
}
\arguments{
\item{t}{a bimets timeseries}
}
\value{
a String representation of a timeseries as a JSON object
}
\description{
Converts to JSON a timeseries
}

