% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{to_list}
\alias{to_list}
\title{converts a timeseries to a list representation}
\usage{
to_list(t)
}
\arguments{
\item{t}{a timeseries}
}
\value{
a list object
}
\description{
converts a timeseries to a list representation
}
