% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{.to_xlsx}
\alias{.to_xlsx}
\title{Exporta in xlsx un Dataset}
\usage{
.to_xlsx(x, path, bycol)
}
\arguments{
\item{x}{Istanza di Dataset}

\item{path}{percorso dove esportare i dati}

\item{bycol}{esporta per colonna se `TRUE`, altrimenti per riga}
}
\description{
Exporta in xlsx un Dataset
}
