% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xlsx.r
\docType{methods}
\name{to_xlsx}
\alias{to_xlsx}
\alias{to_xlsx,Dataset,character,logical-method}
\title{Converte in un file XLSX il dataset}
\usage{
to_xlsx(x, path, bycol = T)

\S4method{to_xlsx}{Dataset,character,logical}(x, path, bycol = T)
}
\arguments{
\item{x}{nome del Dataset}

\item{path}{percorso del file}

\item{bycol}{esporta per colonna se `TRUE`, altrimenti per riga}
}
\description{
Converte in un file XLSX il dataset
}
