% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xlsx.r
\name{to_xlsx}
\alias{to_xlsx}
\title{Converte in un file XLSX il dataset}
\usage{
to_xlsx(x, path, bycol)
}
\arguments{
\item{x}{nome del Dataset}

\item{bycol}{esporta per colonna se `TRUE`, altrimenti per riga}

\item{periodo}{percorso del file}
}
\description{
Converte in un file XLSX il dataset
}
