% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tsRead_nativo}
\alias{tsRead_nativo}
\title{tsRead implementato nativamente.}
\arguments{
\item{path}{a path to the CSV file representing a timeseries}
}
\value{
a `ts` object
}
\description{
tsRead implementato nativamente.
}
