% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tsWrite_nativo}
\alias{tsWrite_nativo}
\title{tsWrite nativo}
\arguments{
\item{series}{timeSeries}

\item{path}{percorso del file}
}
\description{
Scrive una timeseries in formato CSV (legacy support)
}
