% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.r
\docType{methods}
\name{union}
\alias{union}
\alias{union,Dataset,Dataset-method}
\title{Fonde due dataset in un unico dataset
Le serie storiche contenute nel secondo \code{y} sovrascrivono le
serie storiche contenute nel primo in caso che l'intersezione dei
nomi dei due dataset sia non nulla.}
\usage{
union(x, y)

\S4method{union}{Dataset,Dataset}(x, y)
}
\arguments{
\item{x}{il primo dataset}

\item{y}{il secondo dataset}
}
\value{
un dataset con l'unione di tutte le serie storiche
}
\description{
Fonde due dataset in un unico dataset
Le serie storiche contenute nel secondo \code{y} sovrascrivono le
serie storiche contenute nel primo in caso che l'intersezione dei
nomi dei due dataset sia non nulla.
}
\seealso{
base::union
}
