% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlist.r
\docType{methods}
\name{URLIST}
\alias{URLIST}
\alias{URLIST,Dataset-method}
\title{Ritorna un data.frame con metadati delle serie}
\usage{
URLIST(x)

\S4method{URLIST}{Dataset}(x)
}
\arguments{
\item{x}{un Dataset}
}
\value{
un dataframe con 4 colonne (nome, start, end, freq)
}
\description{
I metadati ritornati sono start period, end period, freq
}
