% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlist.r
\name{URLIST}
\alias{URLIST}
\title{Ritorna un data.frame con metadati delle serie}
\usage{
URLIST(x)
}
\arguments{
\item{x}{un Dataset}
}
\value{
un dataframe con 4 colonne (nome, start, end, freq)
}
\description{
I metadati ritornati sono start period, end period, freq
}
