% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{window}
\alias{window}
\title{The 'window' generic}
\usage{
window(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{see params for `stats::window`}
}
\value{
An object of the same class as `x`, possibly modified
}
\description{
The 'window' generic
}
\seealso{
`stats::window()`
}
