% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.r
\name{as.data.frame.Dataset}
\alias{as.data.frame.Dataset}
\title{converte il Dataset in un dataframe}
\usage{
\method{as.data.frame}{Dataset}(x, ...)
}
\description{
Ogni riga del dataframe contiene NOME della serie
Osservazione, anno, periodo, frequenza e data come
timestamp dell'osservazione
}
