% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{as.list}
\alias{as.list}
\title{Ritorna il dataset come una \code{list} di serie storiche}
\usage{
\S4method{as.list}{Dataset}(x)
}
\arguments{
\item{x}{il Dataset}
}
\value{
la \code{list} contenente le serie storiche
}
\description{
Ritorna il dataset come una \code{list} di serie storiche
}
\author{
Giuseppe Acito
}

