% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{class}
\name{Dataset}
\alias{Dataset}
\alias{Dataset-class}
\title{Dataset OOP}
\usage{
Dataset(url, ids)
}
\description{
Classe contenitore di dati (serie storiche)
}
\section{Slots}{

\describe{
\item{\code{data}}{hash containing data}

\item{\code{url}}{path where data is}
}}

\author{
Giuseppe Acito
}
