% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{dataset_intersect}
\alias{dataset_intersect}
\alias{>.dataset}
\title{Set Intersection}
\usage{
dataset_intersect(a, b)

\method{>}{dataset}(a, b)
}
\arguments{
\item{a}{A dataset object (left operand).}

\item{b}{A dataset object (right operand, used as filter).}
}
\value{
A dataset containing values from \code{a} where corresponding values in \code{b} are not \code{NA}.
}
\description{
Returns values from dataset \code{a} that are also present in dataset \code{b}.
A value is considered present if it is not \code{NA}. For each cell identified by
(row_id, col_id), the value from \code{a} is kept only if the corresponding cell in \code{b} is not \code{NA}.
Dataset \code{b} acts as a filter determining which values from \code{a} appear in the result.
}
\details{
If the ID columns of \code{a} and \code{b} do not match.
}
\keyword{internal}
