% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{dataset_minus}
\alias{dataset_minus}
\alias{-.dataset}
\title{Set Difference}
\usage{
dataset_minus(a, b)

\method{-}{dataset}(a, b)
}
\arguments{
\item{a}{A dataset object (minuend).}

\item{b}{A dataset object (subtrahend).}
}
\value{
A dataset containing values from \code{a} where corresponding values in \code{b} are \code{NA}.
}
\description{
Returns values from dataset \code{a} that are not present in dataset \code{b}.
A value is considered present if it is not \code{NA}. For each cell identified by
(row_id, col_id), the value from \code{a} is kept only if the corresponding cell in \code{b} is \code{NA}.
}
\details{
If the ID columns of \code{a} and \code{b} do not match.
}
\keyword{internal}
