% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{djoin}
\alias{djoin}
\title{Joina i due dataset \code{x} e \code{historic} alla data (\code{date})
specificata.}
\usage{
\S4method{djoin}{Dataset,Dataset}(historic, x, date)
}
\arguments{
\item{historic}{il dataset storico}

\item{x}{il dataset nuovo}

\item{date}{una periodo di join}
}
\value{
Un dataset joinato
}
\description{
Joina i due dataset \code{x} e \code{historic} alla data (\code{date})
specificata.
}
\author{
Giuseppe Acito
}
