% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{saveDataset}
\alias{saveDataset}
\title{Salva un Dataset}
\usage{
\S4method{saveDataset}{Dataset,character,logical}(x, path, as.csv = FALSE)
}
\arguments{
\item{x}{un Dataset}

\item{path}{percorso dove salvare il Dataset}

\item{as.csv}{Salva Dataset come CSV
files csv (1 per serie storica)}
}
\description{
Salva un Dataset.
}
\author{
Giuseppe Acito
}
\seealso{
\code{saveGraph}
}

