% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{saveDataset}
\alias{saveDataset}
\title{Salva un Dataset}
\usage{
\S4method{saveDataset}{Dataset,character,logical,logical}(x, path,
  as.csv = FALSE, as.grafo = FALSE)
}
\arguments{
\item{x}{un Dataset}

\item{path}{percorso dove salvare il Dataset}

\item{as.csv}{Salva Dataset come CSV
files csv (1 per serie storica)}

\item{as.grafo}{Salva Dataset con la struttura su filesystem compatible con
il Grafo (implica as.csv == TRUE)}
}
\description{
Salva un Dataset.
}
\author{
Giuseppe Acito
}
\seealso{
\code{saveGraph}
}

