% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{to_xlsx}
\alias{to_xlsx}
\title{Allunga tutte le serie del Dataset.}
\usage{
to_xlsx(x, path, bycol)
}
\arguments{
\item{x}{nome del Dataset}

\item{bycol}{esporta per colonna se `TRUE`, altrimenti per riga}

\item{periodo}{percorso del file}
}
\description{
La funzione segue la seguente convezione:
   - se la serie termina in `C` o `S` e' una condistenza e va allungata
     con il medesimo valore
   - altrimenti e' un flusso e va allungata con zero
}

