% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{window.Dataset}
\alias{window.Dataset}
\alias{window,Dataset-method}
\alias{window,Dataset,ANY-method}
\title{Esegue il `window` sul Dataset}
\usage{
window.Dataset(x, ...)

\S4method{window}{Dataset}(x, ...)
}
\arguments{
\item{x}{Dataset su cui applicare il window}

\item{...}{altri parametri da passare a stats::window}
}
\description{
Esegue il `window` sul Dataset
}
